+++ sklearn/ensemble/tests/test_gradient_boosting.py
@@ -62,6 +62,9 @@ def test_classification_toy():
         assert np.any(deviance_decrease >= 0.0), \
             "Train deviance does not monotonically decrease."
 
+        leaves = clf.apply(X)
+        assert leaves.shape == (X.shape[0], clf.n_estimators, clf.n_classes)
+
 
 def test_parameter_checks():
     # Check input parameter validation.