+++ doc/whats_new.rst
@@ -142,6 +142,9 @@ Enhancements
      such that the split that maximizes this value also maximizes the impurity
      improvement. By `Arnaud Joly`_, `Jacob Schreiber`_ and `Gilles Louppe`_
 
+   - :class:'ensemble.GradientBoostingRegressor/Classifier' now exposes
+     an ``apply`` method for retrieving the leaf indices samples are
+     predicted as. By `Jacob Schreiber`_.
 
 Bug fixes
 .........